#11 Refactor GroceryItem
