+++ frontend/src/components/GroceryItem.module.scss
@@ -23,19 +23,3 @@
     width: 30px;
     height: 30px;
 }
-
-.icon {
-    & > * {
-        fill: food.$primary-200;
-    }
-
-    &:hover {
-        & > * {
-            fill: food.$primary-100;
-        }
-    }
-
-    cursor: pointer;
-    width: 35px;
-    height: 35px;
-}
