+++ frontend/src/components/KitchenPage/MealCreator.module.scss
@@ -206,3 +206,19 @@
         justify-content: space-between;
     }
 }
+
+.icon {
+    & > * {
+        fill: food.$primary-200;
+    }
+
+    &:hover {
+        & > * {
+            fill: food.$primary-100;
+        }
+    }
+
+    cursor: pointer;
+    width: 35px;
+    height: 35px;
+}
