+++ frontend/src/components/common/SearchField.tsx
@@ -1,11 +1,17 @@
 import classNames from 'classnames'
-import React, { useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { Button } from './Button'
 import styles from './SearchField.module.scss'
 import { SearchOutlined, CloseOutlined } from '@ant-design/icons'
-import { searchGroceries } from '../../api/main'
+import { manageShoppingList, searchGroceries } from '../../api/main'
 import { Grocery } from '../../api/types'
 import { GroceryItem } from '../GroceryItem'
+import { Icon } from '@iconify/react'
+import { useRecoilState, useRecoilValue } from 'recoil'
+import { shoppingListState } from '../../state/shoppinglist'
+import { manageFavoriteList } from '../../api/favorite'
+import { userState } from '../../state/user'
+import { favoritelistState } from '../../state/favoritelist'
 
 interface Props {
     className?: string
@@ -20,6 +26,9 @@ export const SearchField: React.FC<Props> = ({
 }: Props) => {
     const [searchValue, setSearchValue] = useState<string>('')
     const [searchResult, setSearchResult] = useState<Grocery[] | null>(null)
+    const [shoppingList, setShoppingList] = useRecoilState(shoppingListState)
+    const user = useRecoilValue(userState)
+    const [favoriteList, setFavoriteList] = useRecoilState(favoritelistState)
 
     const search = async () => {
         if (searchValue.length > 0) {
@@ -28,6 +37,42 @@ export const SearchField: React.FC<Props> = ({
             }
             const searchResult = await searchGroceries(searchObject)
             setSearchResult(searchResult)
+        }
+    }
+
+    const onShoppingListClick = (grocery: Grocery) => {
+        if (shoppingList) {
+            if (
+                shoppingList.groceries
+                    .map((grocery) => grocery._id)
+                    .indexOf(grocery._id) == -1
+            ) {
+                manageShoppingList('add', grocery._id, shoppingList._id).then(
+                    (shl) => setShoppingList(shl)
+                )
+            } else {
+                manageShoppingList(
+                    'delete',
+                    grocery._id,
+                    shoppingList._id
+                ).then((shl) => setShoppingList(shl))
+            }
+        }
+    }
+
+    const handleGroceryFavorite = (grocery: Grocery) => {
+        if (user) {
+            if (favoriteList.groceries.indexOf(grocery._id) == -1) {
+                manageFavoriteList('add', user.favoritelist, grocery._id).then(
+                    (favoritelist) => setFavoriteList(favoritelist)
+                )
+            } else {
+                manageFavoriteList(
+                    'delete',
+                    user.favoritelist,
+                    grocery._id
+                ).then((favoritelist) => setFavoriteList(favoritelist))
+            }
         }
     }
 
@@ -84,7 +129,47 @@ export const SearchField: React.FC<Props> = ({
             {searchResult && (
                 <>
                     {searchResult.map((grocery: Grocery) => (
-                        <GroceryItem key={grocery.Matvare} grocery={grocery} />
+                        <GroceryItem key={grocery.Matvare} grocery={grocery}>
+                            <>
+                                {favoriteList.groceries.indexOf(grocery._id) !=
+                                -1 ? (
+                                    <Icon
+                                        icon="ant-design:heart-filled"
+                                        className={styles.icon}
+                                        onClick={() =>
+                                            handleGroceryFavorite(grocery)
+                                        }
+                                    />
+                                ) : (
+                                    <Icon
+                                        icon="ant-design:heart-outlined"
+                                        className={styles.icon}
+                                        onClick={() =>
+                                            handleGroceryFavorite(grocery)
+                                        }
+                                    />
+                                )}
+                                {shoppingList?.groceries
+                                    .map((grocery) => grocery._id)
+                                    .indexOf(grocery._id) == -1 ? (
+                                    <Icon
+                                        onClick={() =>
+                                            onShoppingListClick(grocery)
+                                        }
+                                        icon="carbon:shopping-cart-plus"
+                                        className={styles.icon}
+                                    />
+                                ) : (
+                                    <Icon
+                                        onClick={() =>
+                                            onShoppingListClick(grocery)
+                                        }
+                                        icon="carbon:shopping-cart-minus"
+                                        className={styles.icon}
+                                    />
+                                )}
+                            </>
+                        </GroceryItem>
                     ))}
                     <div className={styles.information}>
                         {searchResult.length > 0
